dev-requirements.txt
pylint==2.7.2
pylint-django==2.4.2
parameterized==0.8.1
